added common binder; added ContainerInitOptions
Added a common binder ('ICommonBinder') to allow binding regardless of
container.

Added ContainerInitializationOptions to encapsulate the boolean options
into a single object.

Changed method signatures to use ContainerInitializationOptions so this is
incompatible with previous versions.  I'm assuming no one is using this
library yet, so making interface changes shouldn't be an issue, right?